Include confirm dialog on delete
Changed a couple of things to comply to material guidelines

+++ src/app/pages/children/children.component.ts
@@ -3,8 +3,9 @@ import { ChildrenService } from 'src/app/services/children.service';
 import { Observable } from 'rxjs';
 import { Child } from 'src/app/models/child';
 import * as moment from 'moment';
-import { firestore } from 'firebase';
 import { Router } from '@angular/router';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmActionComponent, ConfirmActionModel } from '../../shared/components/confirm-action/confirm-action.component';
 
 @Component({
   selector: 'app-children',
@@ -16,19 +17,29 @@ export class ChildrenComponent implements OnInit {
   public children$: Observable<Child[]>;
   constructor(
     private router: Router,
-    private childrenService: ChildrenService
+    private childrenService: ChildrenService,
+    private dialog: MatDialog
   ) { }
 
   ngOnInit(): void {
     this.children$ = this.childrenService.List();
   }
 
-  public clickEdit(kid: Child) {
-    this.router.navigateByUrl('edit-child/' + kid.id);
-  }
   public clickRemove(kid: Child) {
-    this.childrenService.Delete(kid.id);
+    const data = new ConfirmActionModel(
+      'Remove Child',
+      `Are you sure you want to delete ${kid.firstName} ${kid.lastName}?`,
+      'Remove');
+    const dialogRef = this.dialog.open<ConfirmActionComponent, ConfirmActionModel, boolean>(
+      ConfirmActionComponent,
+      {data});
+    dialogRef.afterClosed().subscribe(dialogResult => {
+      if (dialogResult) {
+        this.childrenService.Delete(kid.id);
+      }
+    });
   }
+
   public getAge(kid: Child): string {
     const dob = moment(kid.dob);
     const now = moment();
